src/TelegramBot/requirements.txt
aiogram
onnxruntime-gpu